pyproject.toml
[project]
name = "StockSage"
version = "0.1.0"
description = "Hierarchical Multi-Agent System for Top 5 US Stock Selection"
authors = [{ name = "Ranuga Disansa Gamage", email = "go2ranuga@gmail.com" }, { name = "Sasvidu Abesinghe", email = "amsranthul@gmail.com" }, { name = "Yasiru Ravisara Fernando", email = "dyrfernando@gmail.com" }]
requires-python = ">=3.11,<3.12"
dependencies = [
    "aiohttp>=3.11.13",
    "alpha-vantage>=3.0.0",
    "crewai-tools>=0.36.0",
    "crewai[tools]>=0.102.0,<1.0.0",
    "firecrawl-py>=1.6.8",
    "langsmith>=0.1.147",
    "litellm>=1.60.2",
    "lxml>=5.3.1",
    "mkdocs>=1.6.1",
    "nest-asyncio>=1.6.0",
    "plotly>=6.0.0",
    "pydantic==2.9.2",
    "pydantic-core>=2.14.5",
    "pytest-asyncio>=0.25.3",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.0",
    "textblob>=0.19.0",
    "vadersentiment>=3.3.2",
    "yfinance>=0.2.54",
]

[project.scripts]
stocksage = "stocksage.main:run"
run_crew = "stocksage.main:run"
train = "stocksage.main:train"
replay = "stocksage.main:replay"
test = "stocksage.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[dependency-groups]
dev = [
    "bump-pydantic>=0.8.0",
    "pytest>=8.3.5",
]
